# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreatePollRequest:
    properties:
      allow_customer_choice:
        type: boolean
      choices:
        items:
          type: string
        type: array
      is_multi:
        type: boolean
      poll_name:
        type: string
      userID:
        type: string
    required:
    - allow_customer_choice
    - choices
    - is_multi
    - poll_name
    type: object
  dto.PollResponse:
    properties:
      allow_customer_choice:
        type: boolean
      choices:
        items:
          $ref: '#/definitions/dto.PollResultResponse'
        type: array
      id:
        type: integer
      is_multi:
        type: boolean
      poll_name:
        type: string
      qrcode_url:
        type: string
      room_id:
        type: string
    type: object
  dto.PollResultResponse:
    properties:
      choice_name:
        type: string
      id:
        type: integer
      number_vote:
        type: integer
    type: object
  entities.Order:
    properties:
      id:
        type: integer
      total:
        type: number
    type: object
  entities.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  responses.MessageResponse:
    properties:
      message:
        example: example message
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Polltato.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Polltato API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials (email & password)
        in: body
        name: credentials
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Authenticated user and JWT token
          schema:
            additionalProperties: true
            type: object
      summary: Authenticate user and return token
      tags:
      - users
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: User registration payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entities.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.User'
      summary: Register a new user
      tags:
      - users
  /orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Order'
            type: array
      summary: Get all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Order payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/entities.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Order'
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    delete:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.MessageResponse'
      summary: Delete an order by ID
      tags:
      - orders
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Order'
      summary: Get order by ID
      tags:
      - orders
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order update payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/entities.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Order'
      summary: Update an order partially
      tags:
      - orders
  /polls:
    post:
      consumes:
      - application/json
      parameters:
      - description: Poll payload
        in: body
        name: poll
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePollRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PollResponse'
      summary: Create a new poll
      tags:
      - polls
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
      summary: Get all users
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.MessageResponse'
      summary: Delete an user by ID
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entities.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
      summary: Update an user partially
      tags:
      - users
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
      summary: Get currently authenticated user
      tags:
      - users
swagger: "2.0"